Reorganize components for  trips admin view
* Place under /admin
* Fix trips reducer after action renaming
