+++ src/router.jsx
@@ -10,7 +10,7 @@ import Destination from './sections/admin/destinations/Destination';
 import Destinations from './sections/admin/destinations/Destinations';
 import SignUpForm from './sections/signup/SignUpForm';
 import SignupTrip from './sections/signupTrip/SignupTrip';
-import TripRequestsContainer from './sections/tripRequests/TripRequestsContainer';
+import TripRequests from './sections/admin/tripRequests/TripRequests';
 import ConfirmSignUpFormContainer from './sections/signup/ConfirmSignUpFormContainer';
 import NotFound from './commons/NotFound.jsx';
 
@@ -29,7 +29,7 @@ export default(
                 <Route path="/trips/signup" component={SignupTrip} />
                 <Route path="/admin/destinations" component={Destinations} />
                 <Route path="/admin/destinations/:destinationId" component={Destination} />
-                <Route path="/admin/trips" component={TripRequestsContainer} />
+                <Route path="/admin/trips" component={TripRequests} />
                 <Route path="*" component={NotFound} />
             </Route>
         </Route>
